+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All Notable changes to `Spider` will be documented in this file
 
 ## v0.2 - NEXT
+- Updated Driver Interface to include full api
+- First Party Driver for OrientDB (v2.*)
 
 ## v0.1.1 - 4-28-2015
 ### Added