+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All Notable changes to `Spider` will be documented in this file
 ## v0.2 - NEXT
 - Updated Driver Interface to include full api
 - First Party Driver for OrientDB (v2.*)
+- Define which type of return object you want
 
 ## v0.1.1 - 4-28-2015
 ### Added