+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All Notable changes to `Spider` will be documented in this file
 ## v0.2 - NEXT
 - Updated Driver Interface to include full api
 - First Party Driver for OrientDB (v2.*)
-- Define which type of return object you want
+- Register configs inside Connections\Manager
+- Define return object from Connections\Manager or Connection
 
 ## v0.1.1 - 4-28-2015
 ### Added